requirements.txt
# Streamlit Cloud Optimized Requirements - Python 3.13 Compatible
# Essential packages only for dashboard deployment

# Core packages for dashboard
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.3
numpy>=1.24.3

# Computer Vision (essential for dashboard)
opencv-python-headless>=4.8.1.78
pillow>=7.1.0

# Data visualization
matplotlib>=3.7.2
seaborn>=0.12.2

# Configuration and utilities
pyyaml>=6.0.1
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3

# Optional packages (dashboard will work without these)
# scikit-learn>=1.3.0
# ultralytics>=8.0.0
# scipy>=1.11.1
# loguru>=0.7.0
# psutil>=5.9.5
# python-dotenv>=1.0.0
# imageio>=2.31.1
# imageio-ffmpeg>=0.4.8
# tqdm>=4.65.0
